+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix average statistics computation: ensure failed tasks are not accounted.
+
 ## [2.6.0] - 2023-06-09
 
 ### Added